+++ dsigma/precompute.py
@@ -192,7 +192,7 @@ def precompute(
         Catalog of lenses.
     table_s : astropy.table.Table
         Catalog of sources.
-    bins : numpy.ndarray or u.quantity.Quantity
+    bins : numpy.ndarray or astropy.units.quantity.Quantity
         Bins in radius to use for the stacking. If a numpy array, bins are
         assumed to be in Mpc. If an astropy quantity, one can pass both length
         units, e.g. kpc and Mpc, as well as angular units, i.e. deg and rad.